refactor(story): flatten StoryPhase into components/hooks/types structure
Remove api/ and ui/ wrapper directories. Flatten folder-per-component
into single files. Extract useStoryStream into its own hook. Mirrors
the Projects folder organization.
