+++ frontend/src/features/story/StoryPhase/components/ArtifactPanel.tsx
