+++ frontend/src/features/story/StoryPhase/components/MobileViewToggle.tsx
